Move Table component to components/ui
