--- airbyte-webapp/src/components/Table/index.tsx
@@ -1,3 +1,0 @@
-import Table from "./Table";
-
-export default Table;
